requirements.txt
PyYAML>=6.0.1,<7.0
Flask>=2.3.0,<3.0
requests>=2.31.0,<3.0
geoip2>=4.7.0,<5.0
Flask-Login>=0.6.3,<1.0
Flask-WTF>=1.2.1,<2.0
WTForms>=3.1.1,<4.0
Flask-SocketIO>=5.3.0,<6.0
python-socketio>=5.8.0,<6.0
bcrypt>=4.0.1,<5.0
